+++ pyppeteer/element_handle.py
@@ -3,106 +3,66 @@
 
 """Element handle module."""
 
-import json
 import logging
 import os.path
-from typing import Any, Dict, Optional, TYPE_CHECKING
-import warnings
+from typing import Any, Dict, TYPE_CHECKING
 
-from pyppeteer import helper
+# from pyppeteer import helper
 from pyppeteer.connection import Session
-from pyppeteer.errors import ElementHandleError, BrowserError
-from pyppeteer.input import Mouse, Touchscreen
+from pyppeteer.execution_context import ExecutionContext, JSHandle
+from pyppeteer.errors import ElementHandleError
 
 if TYPE_CHECKING:
     from pyppeteer.frame_manager import Frame  # noqa: F401
+    # from pyppeteer.page import Page  # noqa: F401
 
 
 logger = logging.getLogger(__name__)
 
 
-class ElementHandle(object):
+class ElementHandle(JSHandle):
     """ElementHandle class."""
 
-    def __init__(self, frame: Any, client: Session, remoteObject: dict,
-                 mouse: Mouse, touchscreen: Touchscreen) -> None:
+    def __init__(self, context: ExecutionContext, client: Session,
+                 remoteObject: dict, page: Any) -> None:
         """Make new element handle object."""
-        self._frame: Frame = frame
+        super().__init__(context, client, remoteObject)
         self._client = client
         self._remoteObject = remoteObject
-        self._mouse = mouse
-        self._touchscreen = touchscreen
+        self._page = page
         self._disposed = False
 
-    def _remoteObjectId(self) -> Optional[str]:
-        return None if self._disposed else self._remoteObject['objectId']
+    def asElement(self) -> 'ElementHandle':
+        """Return as an element handle."""
+        return self
 
-    async def dispose(self) -> None:
-        """Release element handle."""
-        if self._disposed:
-            return
-        self._disposed = True
-        await helper.releaseObject(self._client, self._remoteObject)
+    async def _scrollIntoViewIfNeeded(self) -> None:
+        error = await self.executionContext.evaluate(
+            '''element => {
+                if (!element.ownerDocument.contains(element))
+                    return 'Node is detached from document';
+                if (element.nodeType !== Node.ELEMENT_NODE)
+                    return 'Node is not of type HTMLElement';
+                element.scrollIntoViewIfNeeded();
+                return false;
+            }''', self)
+        if error:
+            raise ElementHandleError(error)
 
-    async def evaluate(self, pageFunction: str, *args: Any) -> Any:
-        """[Deprecated] Evaluate the pageFunction on browser."""
-        deprecation_msg = (
-            'ElementHandle.evaluate is dropped in puppeteer. '
-            'Use Page.evaluate(..., ElementHandle) instead.'
-        )
-        logger.warning('[DEPRECATED] ' + deprecation_msg)
-        warnings.warn(deprecation_msg, DeprecationWarning)
-
-        if self._disposed:
-            raise ElementHandleError('ElementHandle is disposed!')
-        _args = ['this']
-        _args.extend(json.dumps(x) for x in args)
-        stringifiedArgs = ','.join(_args)
-        functionDeclaration = f'''
-function() {{ return ({pageFunction})({stringifiedArgs}) }}
-'''
-        objectId = self._remoteObject.get('objectId')
-        obj = await self._client.send(
-            'Runtime.callFunctionOn', {
-                'objectId': objectId,
-                'functionDeclaration': functionDeclaration,
-                'returnByValue': False,
-                'awaitPromise': True,
-            }
-        )
-        exceptionDetails = obj.get('exceptionDetails', dict())
-        remoteObject = obj.get('result', dict())
-        if exceptionDetails:
-            raise BrowserError(
-                'Evaluation failed: ' +
-                helper.getExceptionMessage(exceptionDetails)
-            )
-        return await helper.serializeRemoteObject(self._client, remoteObject)
-
-    async def _visibleCenter(self) -> Dict[str, int]:
-        center = await self._frame.evaluate('''
-element => {
-    if (!element.ownerDocument.contains(element))
-        return null;
-    element.scrollIntoViewIfNeeded();
-    let rect = element.getBoundingClientRect();
-    return {
-        x: (Math.max(rect.left, 0) + Math.min(rect.right, window.innerWidth)) / 2,
-        y: (Math.max(rect.top, 0) + Math.min(rect.bottom, window.innerHeight)) / 2
-    };
-}
-        ''', self)  # noqa: E501
-        if not center:
-            # raise Exception('No node found for selector: ' + selector)
-            raise BrowserError('No node found for selector: ')
-        return center
+    async def _visibleCenter(self) -> Dict[str, float]:
+        await self._scrollIntoViewIfNeeded()
+        box = await self.boundingBox()
+        return {
+            'x': box['x'] + box['width'] / 2,
+            'y': box['y'] + box['height'] / 2,
+        }
 
     async def hover(self) -> None:
         """Move mouse over this element."""
         obj = await self._visibleCenter()
         x = obj.get('x', 0)
         y = obj.get('y', 0)
-        await self._mouse.move(x, y)
+        await self._page.mouse.move(x, y)
 
     async def click(self, options: dict = None, **kwargs: Any) -> None:
         """Click this element."""
@@ -112,7 +72,7 @@ element => {
         if options is None:
             options = dict()
         options.update(kwargs)
-        await self._mouse.click(x, y, options)
+        await self._page.mouse.click(x, y, options)
 
     async def uploadFile(self, *filePaths: str) -> dict:
         """Upload files."""
@@ -123,20 +83,54 @@ element => {
             {'objectId': objectId, 'files': files}
         )
 
-    async def attribute(self, key: str) -> str:
-        """[Deprecated] Get attribute value of the `key` of this element."""
-        deprecation_msg = (
-            'ElementHandle.attribute is dropped in puppeteer. '
-            'Use Page.querySelectorEval or Page.Jeval instead.'
-        )
-        logger.warning('[DEPRECATED]' + deprecation_msg)
-        warnings.warn(deprecation_msg, DeprecationWarning)
-        return await self.evaluate(
-            '(element, key) => element.getAttribute(key)', key)
-
     async def tap(self) -> None:
         """Tap this element."""
         center = await self._visibleCenter()
         x = center.get('x', 0)
         y = center.get('y', 0)
-        await self._touchscreen.tap(x, y)
+        await self._page.touchscreen.tap(x, y)
+
+    async def focus(self) -> None:
+        """Focus this element."""
+        await self.executionContext.evaluate(
+            'element => element.focus()', self)
+
+    async def type(self, text: str, options: Dict = None) -> None:
+        """Type characters on this element."""
+        if options is None:
+            options = {}
+        await self.focus()
+        await self._page.keyboard.type(text, options)
+
+    async def press(self, key: str, options: Dict = None) -> None:
+        """Type key on this element."""
+        if options is None:
+            options = {}
+        await self.focus()
+        await self._page.keyboard.press(key, options)
+
+    async def boundingBox(self) -> Dict[str, float]:
+        """Return bounding box size of this node."""
+        _obj = await self._client.send('DOM.getBoxModel', {
+            'objectId': self._remoteObject.get('objectId'),
+        })
+        model = _obj.get('model')
+        if not model:
+            raise ElementHandleError('node is detached from document')
+
+        quad = model['border']
+        x = min(quad[0], quad[2], quad[4], quad[6])
+        y = min(quad[1], quad[3], quad[5], quad[7])
+        width = max(quad[0], quad[2], quad[4], quad[6]) - x
+        height = max(quad[1], quad[3], quad[5], quad[7]) - y
+        return {'x': x, 'y': y, 'width': width, 'height': height}
+
+    async def screenshot(self, options: Dict = None) -> bytes:
+        """Take a screenshot of this element."""
+        if options is None:
+            options = {}
+        await self._scrollIntoViewIfNeeded()
+        boundingBox = await self.boundingBox()
+        opt = {'clip': boundingBox}
+        opt.update(options)
+        return await self._page.screenshot(opt)